# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/sales": {
      "get": {
        "operationId": "sales_list",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stylistId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "ISSUED",
                "PARTIALLY_PAID",
                "PAID",
                "OVERDUE",
                "VOID"
              ]
            }
          },
          {
            "name": "method",
            "required": false,
            "in": "query",
            "description": "Cobradas con este método",
            "schema": {
              "type": "string",
              "enum": [
                "CASH",
                "CARD",
                "TRANSFER",
                "BIZUM",
                "GIFT_CARD",
                "VOUCHER",
                "OTHER"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoicePageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: invoices.read"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buscar facturas",
        "tags": [
          "Ventas y facturación"
        ]
      },
      "post": {
        "description": "Si alguna línea lleva descuento, exige además el permiso invoices.discount.",
        "operationId": "sales_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceEnvelopeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: invoices.create + payments.create"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Registra una venta: emite la factura, cobra y descuenta existencias por FEFO",
        "tags": [
          "Ventas y facturación"
        ]
      }
    },
    "/api/v1/sales/summary": {
      "get": {
        "operationId": "sales_summary",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-09-29T06:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-09-30T05:59:59.999Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesSummaryEnvelopeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: invoices.read"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cuadre de un periodo: lo facturado y, por método, lo cobrado y lo devuelto",
        "tags": [
          "Ventas y facturación"
        ]
      }
    },
    "/api/v1/sales/{id}": {
      "get": {
        "operationId": "sales_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceEnvelopeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: invoices.read"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar una factura, con sus cobros",
        "tags": [
          "Ventas y facturación"
        ]
      }
    },
    "/api/v1/sales/{id}/void": {
      "post": {
        "description": "El efectivo sale de la caja abierta: sin caja abierta no se anula una venta cobrada en efectivo. Los cobros con tarjeta o transferencia quedan devueltos y se reembolsan por su medio. El stock vuelve a los mismos lotes.",
        "operationId": "sales_void",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoidInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceEnvelopeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: invoices.void"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Anula una venta devolviendo sus cobros y su género (ADR-0020)",
        "tags": [
          "Ventas y facturación"
        ]
      }
    },
...
  },
  "info": {
    "title": "AppSalonBelleza API",
...
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desarrollo local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token obtenido en POST /api/v1/auth/login"
      }
    },
    "schemas": {
...
      "PageMetaResponse": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 137,
            "description": "Total de elementos que cumplen el filtro"
          },
          "totalPages": {
            "type": "number",
            "example": 7
          },
          "hasNext": {
            "type": "boolean",
            "example": true
          },
          "hasPrevious": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages",
          "hasNext",
          "hasPrevious"
        ]
      },
...
      "InvoiceLineResponse": {
        "type": "object",
        "properties": {
          "stylistName": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "productId": {
            "type": "string",
            "nullable": true
          },
          "serviceId": {
            "type": "string",
            "nullable": true
          },
          "stylistId": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "string"
          },
          "unitPrice": {
            "type": "string"
          },
          "discountAmount": {
            "type": "string"
          },
          "taxRate": {
            "type": "string"
          },
          "taxAmount": {
            "type": "string"
          },
          "lineSubtotal": {
            "type": "string"
          },
          "lineTotal": {
            "type": "string"
          },
          "commissionAmount": {
            "type": "string"
          }
        },
        "required": [
          "stylistName",
          "id",
          "kind",
          "productId",
          "serviceId",
          "stylistId",
          "description",
          "quantity",
          "unitPrice",
          "discountAmount",
          "taxRate",
          "taxAmount",
          "lineSubtotal",
          "lineTotal",
          "commissionAmount"
        ]
      },
      "PaymentResponse": {
        "type": "object",
        "properties": {
          "refundedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "refundReason": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "refundedAmount": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "reference": {
            "type": "string",
            "nullable": true
          },
          "receivedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "refundedAt",
          "refundReason",
          "id",
          "method",
          "status",
          "amount",
          "refundedAmount",
          "currency",
          "reference",
          "receivedAt"
        ]
      },
      "InvoiceResponse": {
        "type": "object",
        "properties": {
          "clientName": {
            "type": "string",
            "nullable": true
          },
          "createdByName": {
            "type": "string",
            "nullable": true,
            "description": "Quién registró la venta: la cajera que la cobró."
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineResponse"
            }
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentResponse"
            }
          },
          "id": {
            "type": "string"
          },
          "number": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "clientId": {
            "type": "string",
            "nullable": true
          },
          "appointmentId": {
            "type": "string",
            "nullable": true
          },
          "issuedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "dueAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "paidAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "voidedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "voidReason": {
            "type": "string",
            "nullable": true
          },
          "subtotal": {
            "type": "string"
          },
          "discountTotal": {
            "type": "string"
          },
          "taxTotal": {
            "type": "string"
          },
          "total": {
            "type": "string"
          },
          "paidTotal": {
            "type": "string"
          },
          "balanceDue": {
            "type": "string"
          },
          "commissionTotal": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "clientName",
          "createdByName",
          "lines",
          "id",
          "number",
          "status",
          "clientId",
          "appointmentId",
          "issuedAt",
          "dueAt",
          "paidAt",
          "voidedAt",
          "voidReason",
          "subtotal",
          "discountTotal",
          "taxTotal",
          "total",
          "paidTotal",
          "balanceDue",
          "commissionTotal",
          "currency",
          "notes"
        ]
      },
      "InvoicePageResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceResponse"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaResponse"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "PaymentMethodTotalsResponse": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string"
          },
          "received": {
            "type": "string"
          },
          "refunded": {
            "type": "string"
          },
          "net": {
            "type": "string",
            "description": "Cobrado menos devuelto: lo que debe haber quedado por ese medio."
          }
        },
        "required": [
          "method",
          "received",
          "refunded",
          "net"
        ]
      },
      "SalesSummaryResponse": {
        "type": "object",
        "properties": {
          "byMethod": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentMethodTotalsResponse"
            }
          },
          "count": {
            "type": "number"
          },
          "total": {
            "type": "string"
          },
          "discountTotal": {
            "type": "string"
          },
          "voidCount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          }
        },
        "required": [
          "byMethod",
          "count",
          "total",
          "discountTotal",
          "voidCount",
          "currency"
        ]
      },
      "SalesSummaryEnvelopeResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SalesSummaryResponse"
          }
        },
        "required": [
          "data"
        ]
      },
      "InvoiceEnvelopeResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/InvoiceResponse"
          }
        },
        "required": [
          "data"
        ]
      },
      "SaleLineDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "PRODUCT",
              "SERVICE"
            ]
          },
          "itemId": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "example": 1
          },
          "stylistId": {
            "type": "string"
          },
          "discountAmount": {
            "type": "number",
            "example": 5,
            "description": "Descuento en importe sobre la línea"
          }
        },
        "required": [
          "kind",
          "itemId",
          "quantity"
        ]
      },
      "SalePaymentDto": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "CASH",
              "CARD",
              "TRANSFER",
              "BIZUM",
              "GIFT_CARD",
              "VOUCHER",
              "OTHER"
            ]
          },
          "amount": {
            "type": "number",
            "example": 45.5
          },
          "reference": {
            "type": "string"
          }
        },
        "required": [
          "method",
          "amount"
        ]
      },
      "CreateSaleDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string"
          },
          "appointmentId": {
            "type": "string"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SaleLineDto"
            }
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SalePaymentDto"
            }
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "lines",
          "payments"
        ]
      },
      "VoidInvoiceDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Cobrada por error a la clienta equivocada"
          }
        },
        "required": [
          "reason"
        ]
      },
...
    }
  }
}